move extrinsic composition from block importer to indirect calls executor
