+++ core/parentchain/block-importer/Cargo.toml
@@ -32,9 +32,6 @@ thiserror = { version = "1.0", optional = true }
 codec  = { package = "parity-scale-codec", version = "3.0.0", default-features = false, features = ["derive"] }
 log = { version = "0.4", default-features = false }
 sp-runtime = { default-features = false, git = "https://github.com/paritytech/substrate.git", branch = "polkadot-v0.9.24" }
-#beefy-merkle-tree = { default-features = false, git = "https://github.com/paritytech/substrate.git", branch = "polkadot-v0.9.24", features = "keccak" }
-#remove as soon as we can import beefy-merkle-tree:
-tiny-keccak = { version = "2.0.2", features = ["keccak"] }
 
 [features]
 default = ["std"]
