+++ core/parentchain/indirect-calls-executor/Cargo.toml
@@ -33,6 +33,9 @@ codec  = { package = "parity-scale-codec", version = "3.0.0", default-features =
 log = { version = "0.4", default-features = false }
 sp-core = { default-features = false, features = ["full_crypto"], git = "https://github.com/paritytech/substrate.git", branch = "polkadot-v0.9.24" }
 sp-runtime = { default-features = false, git = "https://github.com/paritytech/substrate.git", branch = "polkadot-v0.9.24" }
+#beefy-merkle-tree = { default-features = false, git = "https://github.com/paritytech/substrate.git", branch = "polkadot-v0.9.24", features = "keccak" }
+#remove as soon as we can import beefy-merkle-tree:
+tiny-keccak = { version = "2.0.2", features = ["keccak"] }
 
 # scs/integritee
 itp-types = { default-features = false, git = "https://github.com/integritee-network/pallets.git", branch = "master" }
